scl: add scl lock
Add the implementation of the Scheduler Cooperative Lock (SCL) to LiTL.
The original paper of SCL can be found here:
https://research.cs.wisc.edu/adsl/Publications/eurosys20-scl.pdf

This commit includes the following changes
- Makefile.config: add scl_original to Makefile
- src/interpose.c: add scl.h header
- include/scl.h: header file for scl implementation
- src/scl.h: source file for scl implementation

Notice that the current implementation only includes the basic
*_mutex_lock and *_mutex_unlock methods and skipped the cv
implementation.

The implementation largely borrows the original implementation by the
paper's author. The original implementation of SCL can be found here:
https://github.com/scheduler-cooperative-locks/proportional-share

Signed-off-by: YuChen Qian <yuchenqian96@gmail.com>
Reviewed-by:
Cc-Team:
CR: