scl: user-defined CYCLE_PER_US and FAIRLOCK_GRANULARITY
The CYCLE_PER_US should be defined with the CPU cycle of the machine,
and pass in through the Makefile. Similarly, FAIRLOCK_GRANULARITY is
used extensively in the benchmark to define various sizes of lock
slices.

This commit changes the corresponding Makefiles and the source file to
ask the user to define these two macros.

As an example
> `make -j4 CYCLE_PER_US=2400 FAIRLOCK_GRANULARITY=4800000`

Signed-off-by: YuChen Qian <yuchenqian96@gmail.com>
Reviewed-by:
Cc-Team:
CR: